created portfolios component. fix porfolios model on server

+++ node/project/react/src/components/Registration/Registration.tsx
@@ -5,41 +5,58 @@ import cls from '../Login/Login.module.css';
 
 import api from '../../api/axios';
 import { useNavigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import type { RootState } from '../../state';
+import { setToken } from '../../state/slices/authSlice';
 
 interface RegistrationProps { }
 
 const Registration: FunctionComponent<RegistrationProps> = () => {
-  const [loginError, setloginError] = useState<boolean>(false);
+  
+  const [emailError, setemailError] = useState<boolean>(false);
+
   const navigate = useNavigate()
 
-  const user = useSelector<RootState>(s => s.auth)
+  const token = useSelector((s: RootState) => s.auth.accessToken)
+
+  const dispatch = useDispatch()
 
   useEffect(() => {
-    if (user) {
+    if (token) {
       navigate('../dashboard')
     }
-  }, [])
+  }, [token])
 
   const formik = useFormik({
     initialValues: {
       email: '',
       name: '',
       password: '',
+      confirmPassword: '',
     },
     validationSchema: Yup.object({
-      email: Yup.string().email('Невірний email').required('Обов’язково'),
-      name: Yup.string().min(1).max(64).required('Обов’язково'),
-      password: Yup.string().min(6, 'Мінімум 6 символів').max(64, 'Максимум 64 символи').required('Обов’язково'),
+      email: Yup.string().email().required(),
+      name: Yup.string().min(1).max(64).required(),
+      password: Yup.string().min(6).max(64).required(),
+      confirmPassword: Yup.string()
+        .oneOf([Yup.ref('password')], 'confirm password is not same as password')
+        .required(),
     }),
     onSubmit: (values) => {
-      api.post('/user/register', { ...values }, {
-          withCredentials: true
+      const { confirmPassword, ...dataToSend } = values;
+      api.post('/user/register', { ...dataToSend }, {
+        withCredentials: true
       }).then(res => {
         localStorage.setItem('accessToken', res.data.accessToken)
+        dispatch(setToken(res.data.accessToken))
 
-
+      }).catch((err)=>{
+        if(err.status === 409) {
+          setemailError(true)
+        }
+        else{
+          alert('something wrong')
+        }
       })
     },
   });
@@ -48,8 +65,8 @@ const Registration: FunctionComponent<RegistrationProps> = () => {
     <div className={cls.container}>
       <form onSubmit={formik.handleSubmit} className={cls.form}>
 
-        <h2 className={cls.title}>Вхід до акаунту</h2>
-
+        <h2 className={cls.title}>Registration</h2>
+        {emailError && <div className={cls.error}>email is already in use</div>}
         <div className={cls.field}>
           <label htmlFor="email" className={cls.label}>Email</label>
           <input
@@ -77,7 +94,7 @@ const Registration: FunctionComponent<RegistrationProps> = () => {
         </div>
 
         <div className={cls.field}>
-          <label htmlFor="password" className={cls.label}>Пароль</label>
+          <label htmlFor="password" className={cls.label}>password</label>
           <input
             id="password"
             type="password"
@@ -89,10 +106,23 @@ const Registration: FunctionComponent<RegistrationProps> = () => {
           )}
         </div>
 
+        <div className={cls.field}>
+          <label htmlFor="confirmPassword" className={cls.label}>Confirm Password</label>
+          <input
+            id="confirmPassword"
+            type="password"
+            {...formik.getFieldProps('confirmPassword')}
+            className={cls.input}
+          />
+          {formik.touched.confirmPassword && formik.errors.confirmPassword && (
+            <div className={cls.error}>{formik.errors.confirmPassword}</div>
+          )}
+        </div>
+
         <button type="submit" className={cls.button}>
           Registration
         </button>
-        <button className={cls.registration} type='button' onClick={() => { navigate('../login') }}>I have allready an account</button>
+        <button className={cls.registration} type='button' onClick={() => { navigate('../login') }}>I have already an account</button>
       </form>
     </div>
   );
